requirements.txt
fastapi~=0.115.12
pydantic~=2.11.5
openai~=1.82.0
python-dotenv~=1.1.0
pdfkit~=1.0.0
python-docx~=1.1.2
cloudinary~=1.44.0
requests~=2.32.3
uvicorn~=0.34.3
python-multipart~=0.0.20
h11~=0.16.0
pip~=21.2.4
attrs~=25.3.0
distro~=1.9.0
wheel~=0.37.0
protobuf~=6.31.1
lxml~=5.4.0
future~=0.18.2
idna~=3.10
multidict~=6.4.4
propcache~=0.3.1
exceptiongroup~=1.3.0
sniffio~=1.3.1
click~=8.1.8
httpcore~=1.0.9
httpx~=0.28.1
jiter~=0.10.0
anyio~=4.9.0
numpy~=2.0.2
twilio~=9.6.2
PyJWT~=2.10.1
aiohttp~=3.12.9
aiohttp-retry~=2.9.1
yarl~=1.20.0
async-timeout~=5.0.1
aiosignal~=1.3.2
frozenlist~=1.6.2
aiohappyeyeballs~=2.6.1
certifi~=2025.4.26
starlette~=0.46.2
urllib3~=2.4.0
annotated-types~=0.7.0
charset-normalizer~=3.4.2
itsdangerous~=2.2.0
PyYAML~=6.0.2
six~=1.15.0
setuptools~=58.0.4
packaging~=25.0
dotenv~=0.9.9